cello-v2/src/api-engine/requirements.txt
# Driver PostgreSQL obrigatório para Django
psycopg2-binary==2.9.9
asgiref==3.6.0
attrs==20.3.0
cachetools==4.1.1
celery==5.2.6
certifi==2024.7.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==4.2.16
django-allauth==0.55.0
django-cors-headers==3.5.0
django-extensions==3.2.3
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==4.2.0
drf-yasg==1.21.7
flex==6.14.1
google-auth==1.23.0
holdup==1.8.0
uwsgi==2.0.27
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonpointer==2.0
jsonschema==3.2.0
kubernetes==11.0.0
MarkupSafe==2.0.1
oauthlib==3.1.0
packaging==21.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
PyYAML==5.3.1
redis==4.4.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
PyYAML==5.3.1
redis==4.4.4
requests==2.32.2
requests-oauthlib==1.3.0
rfc3987==1.3.8
rsa==4.7
validate-email==1.3
websocket-client==0.57.0
fqdn===1.5.1
validate-email==1.3
websocket-client==0.57.0
fqdn===1.5.1
